Don't forget to pass an exp key to the token, otherwise the token will never expire!
Make sure a bad password fails.
Check for a 401 status code. Copy the method name and go run that test:
bin/phpunit --filter testPOSTTokenInvalidCredentials

It should pass... but it doesn't! Interesting. Look at this: it definitely doesn't return the token... it redirected us to /login. We are getting kicked out of the controller, but this is not how we want our API error responses to work.