First: getCredentials(). Our job is to read the Authorization header and return the token - if any - that's being passed. To help with this, we can use an object from the JWT bundle we installed earlier. If there is no token, return null. This will cause authentication to stop. Not fail, just stop trying to authenticate the user via this method. If there is a token, return it!
getUser()

Next, Symfony will call getUser() and pass this token string as the $ credentials argument. Our job here is to use that token to find the user it relates to.

And this is where JSON web tokens really shine. Because if we simply decode the token, it will contain the username. Then, we can just look it up in the database.

$ data is now an array of whatever information we originally put into the token. Fundamentally, this works just like a normal json_decode, except that the library is also checking to make sure that the contents of our token weren't changed. It does this by using our private key. This guarantees that nobody has changed the username to some other username because they're a jerk. Encryption is amazing.

It also checks the token's expiration: our tokens last 1 hour because that's what we setup in config.yml

checkCredentials()

If the user is not found, this will return null and authentication will fail. But if a user is found, then Symfony finally calls checkCredentials(). Just return true. There's no password or anything else we need to check at this point.

Skip onAuthenticationFailure() for now. And for onAuthenticationSuccess(), purposefully do nothing. We want the authenticated request to continue to the controller so we can do our normal work.

In supportsRememberMe() - this doesn't apply to us - so return false.

And keep start() blank for another minute. With just getCredentials() and getUser() filled in, our authenticator is ready to go.